Signal handling: Wait for all workers to terminate
The promise returned from `Promise.all()` immediately rejects if any
of the underlying promises reject, even if some of the other promises
have not yet settled. Catch each rejection to give the other promises
an opportunity to resolve.

`Promise.allSettled()` could be used instead, but that's a relatively
new function that was added in Node.js v12.9.0.